feat: Create helpers module - Add void? check to hot_dog.bun
