--- constants.rb
@@ -1,10 +1,0 @@
-DIVIDER = "#{'-' * 80}"
-PROMPT = "--> "
-
-OPTION_ERROR = "Please enter a valid option"
-TYPE_ERROR = "Please enter a number"
-
-MENU_OPTIONS = ["quit", "place an order"]
-HOT_DOG_OPTIONS = ["polish", "beef", "jalepano", "tofu", "mystery meat"]
-BUN_OPTIONS = ["hold the bun", "whole wheat", "gluten-free", "tortilla"]
-CONDIMENT_OPTIONS = ["none", "ketchup", "mustard", "relish", "horseradish", "onions", "sauerkraut"]
